--- src/@batch-flask/ui/file/file-viewer/file-viewer-container/file-content/file-content.html
@@ -1,2 +1,0 @@
-<template #viewerContainer></template>
-<bl-file-too-large [file]="file" [fileLoader]="fileLoader" *ngIf="fileTooLarge"></bl-file-too-large>
