--- src/@batch-flask/ui/file/file-viewer/file-viewer-container/file-content/file-content.scss
@@ -1,30 +1,0 @@
-@import "app/styles/variables";
-
-bl-file-content {
-    .viewer-container {
-        height: 100%;
-        > * {
-            height: 100%;
-        }
-    }
-    .unknown-file-type {
-        display: flex;
-        align-items: center;
-
-        .title {
-            flex: 1;
-        }
-
-        .options > button {
-            padding: 10px;
-            margin: 5px;
-            cursor: pointer;
-            background: map-get($primary, 300);
-            color: $whitesmoke;
-
-            &:hover {
-                background: map-get($primary, 500);
-            }
-        }
-    }
-}
