--- src/@batch-flask/ui/file/file-viewer/file-viewer-container/file-content/index.ts
@@ -1,1 +1,0 @@
-export * from "./file-content.component";
