+++ src/@batch-flask/ui/file/file-viewer/file-viewer-container/index.ts
@@ -1,2 +1,1 @@
 export * from "./file-viewer-container.component";
-export * from "./file-content";
