+++ src/@batch-flask/ui/file/file-viewer/file-viewer.module.ts
@@ -9,12 +9,11 @@ import { CodeFileViewerComponent } from "./code-file-viewer";
 import { FileDialogViewerComponent } from "./file-dialog-viewer/file-dialog-viewer.component";
 import { FileTooLargeComponent } from "./file-too-large";
 import { FileTypeAssociationService } from "./file-type-association";
-import { FileContentComponent, FileViewerContainerComponent } from "./file-viewer-container";
+import { FileViewerContainerComponent } from "./file-viewer-container";
 import { ImageFileViewerComponent } from "./image-file-viewer";
 import { LogFileViewerComponent } from "./log-file-viewer";
 
 const publicComponents = [
-    FileContentComponent,
     FileViewerContainerComponent,
     FileTooLargeComponent,
     LogFileViewerComponent,
